exlude unusual devices from disk usage alert
Eg, we had a Pagerduty alert because `loop0`, which Ubuntu mounted as
`/snap/core/...` was at 100%. That's normal and not a problem. The
Stackdriver agent already excludes certain things like `tmpfs` and
`udev`, but apparently doesn't know about loop devices.

This commit adds a filter on the device to only devices that match
`sd.*` which should match actual filesystem mounts.