gcp-project: enable VM Manager
See:
https://appsembler.atlassian.net/wiki/spaces/ORANGE/blog/2022/03/17/2459926695/VM+Manager+quick+demo

This will be handy for vulnerability/patch management and should be
enabled on all of our projects. (Not every VM will immediately work
though; the setup also requires that VMs have service agents
associated with them and not all of our existing ones do, but this
will let us start working through them and make sure that it's all set
up for new infrastructure).

To work out what was needed for this, I created a fresh test project
and kept track of exactly which services and attributes had to be
enabled before I could start up an instance and have it work with VM
Manager, so this should be a pretty minimal set of changes.